Splitting the app.py function
