+++ .serverless/requirements
@@ -1,1 +1,1 @@
-/Users/ashokgadepalli/Library/Caches/serverless-python-requirements/6d10664fcaec2f00dffbece140180c2bd132f21dfd6f961385d0ac92ec292e0e_slspyc
+/Users/ashokgadepalli/Library/Caches/serverless-python-requirements/aabf0aff8753592b730099c2818a9f26ff39ba041893e2b441376fac14b5718c_slspyc
