+++ README.md
@@ -9,25 +9,24 @@ Makes scheduling interview slots for the coming week super easy. Pull out your p
 
 You need four things to run the slackbot locally:
 
-1) Google Calendar API credentials
+1) Your own Slack Workspace for bot development
 
-2) Your own Slack Workspace for bot development
+2) A Slack app in the workspace created above
 
-3) A Slack app in the workspace created above
+3) [ngrok](https://ngrok.com/download)
 
-4) ngrok
+🛑 If you just want to play with a slackbot, finish the steps above and use this [repo](https://github.com/ashokgadeking/python-message-menu-example/tree/blocks_messages) instead.
 
-## Download credentials for the Google Calendar API
+4) Google Calendar API credentials
 
-Go to the page below and click on the `Enable Google Calendar API` button. Download the client config `credentials.json` from the dialog box.
-
-https://developers.google.com/calendar/quickstart/python
 
 ## Create a Slack workspace
 
-Click on the link below and use your personal personal email address (not Contino!) to create a development workspace. Setting up the workspace should be super straight-forward.
+Click on the link below and use your personal personal email address (not Contino email!) to create a development workspace. Setting up the workspace should be super straight-forward.
 
 https://slack.com/create#email
+
+https://slack.com/help/articles/206845317-Create-a-Slack-workspace
 
 ## Configure Slack App on the Slack API console
 
@@ -39,9 +38,11 @@ https://slack.com/create#email
 
 4) Click on OAuth & Permissions and scroll down to Scopes (Bot Token Scopes). Click on `Add an Oauth Scope` and add the scopes from the image below.
 
+![screen capture](./images/scopes.png)
+
 Insert bot scopes image
 
-5) Click on `Install App` button on the sidebar and install app to the workspace you created.
+5) Click on `Install App` button on the sidebar and install app to the development workspace you created in the section above.
 
 ## Setup local development environment
 
@@ -53,7 +54,7 @@ export SLACK_VERIFICATION_TOKEN='blurph-blurph'
 export INTERVIEW_AVAIL_CAL='you@email.com'
 ```
 
-2) Install Python 3.7 (figure it out)
+2) Install Python >= 3.7 (figure it out)
 
 3) Install ngrok. Installation instructions: https://ngrok.com/download
 
@@ -61,8 +62,16 @@ export INTERVIEW_AVAIL_CAL='you@email.com'
 
 6) Copy the URL generated by ngrok, either http or https.
 
-7) Paste the URL into the interactive components `Request URL` section. (The form we left empty in step2 in the section above)
+7) Paste the URL into the interactive components `Request URL` section. (The form we left empty in step2 in the section above).
 
 8) Our application listens on `/slack/message_actions` endpoint so we need to add this to the end of the URL we pasted above.
 
-9) We should be good to go now, run `python3 app.py local` and if the flask server has started successfully, You should see a dm from the bot in your new workspace.
+9) Change the service to local in `app.py`. Comment out the call that fetches delegated crdentials and uncomment `service = calendar_api.get_service_local_creds()`.
+
+9) We should be good to go now, run `python app.py` and if the flask server has started successfully, You should see a dm from the bot in your new workspace.
+
+## Download credentials for the Google Calendar API
+
+Go to the page below and click on the `Enable Google Calendar API` button. Download the client config `credentials.json` from the dialog box.
+
+https://developers.google.com/calendar/quickstart/python
